New CI feature

+++ src/bin/neph_bin.cr
@@ -16,7 +16,7 @@ class NephBin
   
   def parse_option!
     OptionParser.parse! do |parser|
-      parser.banner = "Usage: neph [options]"
+      parser.banner = "Basic usage: neph [options]"
 
       parser.on(
         "-j JOB",
@@ -38,12 +38,6 @@ class NephBin
         @quiet = true
       end
 
-      parser.on("-c", "--clean", "Cleaning caches") do
-        log_ln "cleaning caches".colorize.fore(:green).mode(:bold).to_s
-        clean
-        exit 0
-      end
-
       parser.on("-v", "--version", "Show the version") do
         log_ln VERSION
         exit 0
@@ -51,13 +45,53 @@ class NephBin
 
       parser.on("-h", "--help", "Show this help") do
         log_ln parser.to_s
+        log_ln "Action: neph [action]"
+        log_ln "    neph clean     - Cleaning every caches"
+        log_ln "    neph uninstall - Uninstalling neph command"
         exit 0
+      end
+
+      parser.unknown_args do |args|
+        args.each do |arg|
+          case arg
+          when "clean"
+            log_ln "Neph".colorize.fore(:green).mode(:bold).to_s + " is cleaning caches ..."
+            clean
+            exit 0
+          when "uninstall"
+            log_ln "Neph".colorize.fore(:green).mode(:bold).to_s + " will be uninstalled ..."
+            uninstall
+            exit 0
+          end
+        end
       end
     end
   end
 
   def clean
-    FileUtils.rm_rf(NEPH_DIR)
+    FileUtils.rm_rf(NEPH_DIR) if Dir.exists?(NEPH_DIR)
+  end
+
+  def uninstall
+    neph_path = `which neph`.chomp
+    log_ln "Find 'neph' at #{neph_path}"
+
+    loop do
+      log_ln "Are you sure you want to uninstall 'neph'? [Y/n]"
+      
+      if input = STDIN.gets
+        case input
+        when "Y"
+          log_ln "Uninstalling neph..."
+          FileUtils.rm_rf(neph_path)
+          log_ln "Successfully uninstalled, thanks for using 'neph' so far!"
+          break
+        when "n"
+          log_ln "Abort the uninstallation..."
+          break
+        end
+      end
+    end
   end
 
   def exec
